+++ modules/frontend/combiner/trace_by_id_v2.go
@@ -12,7 +12,7 @@ func NewTraceByIDV2(maxBytes int, marshalingFormat string) Combiner {
 	var partialTrace bool
 	gc := &genericCombiner[*tempopb.TraceByIDResponse]{
 		combine: func(partial *tempopb.TraceByIDResponse, _ *tempopb.TraceByIDResponse, _ PipelineResponse) error {
-			if partial.Status == tempopb.TraceByIDResponse_PARTIAL {
+			if partial.Status == tempopb.PartialStatus_PARTIAL {
 				partialTrace = true
 			}
 			_, err := combiner.Consume(partial.Trace)
@@ -30,7 +30,7 @@ func NewTraceByIDV2(maxBytes int, marshalingFormat string) Combiner {
 			resp.Trace = traceResult
 
 			if partialTrace || combiner.IsPartialTrace() {
-				resp.Status = tempopb.TraceByIDResponse_PARTIAL
+				resp.Status = tempopb.PartialStatus_PARTIAL
 				resp.Message = fmt.Sprintf("Trace exceeds maximum size of %d bytes, a partial trace is returned", maxBytes)
 			}